pyproject.toml
[tool.poetry]
name = "packages"
version = "0.1.0"
description = "Requirements for Django project"
authors = ["Baha Dzhalalov <bahadzhalalov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
stripe = "^5.0.0"
psycopg2 = "^2.9.5"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.core.masonry.api"
